+++ providers/databricks/src/airflow/providers/databricks/hooks/databricks.py
@@ -307,6 +307,13 @@ class WarehouseState:
     def __repr__(self) -> str:
         return str(self.__dict__)
 
+    def to_json(self) -> str:
+        return json.dumps(self.__dict__)
+
+    @classmethod
+    def from_json(cls, data: str) -> WarehouseState:
+        return WarehouseState(**json.loads(data))
+
 
 class DatabricksHook(BaseDatabricksHook):
     """
@@ -798,6 +805,15 @@ class DatabricksHook(BaseDatabricksHook):
         :param warehouse_id: ID of the SQL warehouse.
         """
         return WarehouseState(self.get_warehouse(warehouse_id)["state"])
+
+    async def a_get_warehouse_state(self, warehouse_id: str) -> WarehouseState:
+        """
+        Async version of `get_warehouse_state`.
+
+        :param warehouse_id: ID of the SQL warehouse.
+        """
+        response = await self._a_do_api_call(("GET", f"{SQL_WAREHOUSES_ENDPOINT}/{warehouse_id}"))
+        return WarehouseState(response["state"])
 
     def start_warehouse(self, warehouse_id: str) -> None:
         """