+++ providers/databricks/docs/changelog.rst
@@ -55,6 +55,11 @@ Doc-only
 7.18.1
 ......
 
+Features
+~~~~~~~~
+
+* ``Add deferrable mode to Databricks SQL warehouse start and stop operators``
+
 Misc
 ~~~~